fix(ci): coverage shard 8 — split provider_readiness part_001 under 1500-line limit
part_001 grew to 1602 lines, tripping the first-party file-line maintainability
guardrail. Move the launch-preflight payload and all-green readiness tests into a
new part_006 (importing shared helpers from part_001), restoring part_001 to 1387
lines without altering any assertions.
